# contracts/chat/openapi.json
{
  "components": {
    "schemas": {
...
      "ConnectionSeed": {
        "additionalProperties": false,
        "description": "합성 연결의 명시적 준비 입력입니다.",
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "external_conversation_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Conversation Id",
            "type": "string"
          },
          "external_sender_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Sender Id",
            "type": "string"
          },
          "idempotency_supported": {
            "default": true,
            "title": "Idempotency Supported",
            "type": "boolean"
          },
          "lookup_supported": {
            "default": true,
            "title": "Lookup Supported",
            "type": "boolean"
          },
          "operator_user_id": {
            "format": "uuid",
            "title": "Operator User Id",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Run Id",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile",
          "connection_id",
          "external_conversation_id",
          "external_sender_id",
          "operator_user_id"
        ],
        "title": "ConnectionSeed",
        "type": "object"
      },
...
      "DeliveryState": {
        "enum": [
          "pending",
          "sending",
          "accepted",
          "rejected",
          "unknown"
        ],
        "title": "DeliveryState",
        "type": "string"
      },
      "ErrorCode": {
        "enum": [
          "UNAUTHENTICATED",
          "CONVERSATION_NOT_FOUND",
          "IDEMPOTENCY_CONFLICT",
          "INVALID_INPUT",
          "NOT_FOUND",
          "METHOD_NOT_ALLOWED",
          "HTTP_ERROR",
          "INTERNAL_ERROR",
          "DATABASE_BUSY",
          "DATABASE_POOL_TIMEOUT",
          "INSTANCE_MISMATCH",
          "EVENT_IDENTITY_MISMATCH",
          "REGISTRY_UNAVAILABLE",
          "RESYNC_NOT_CONFIRMED"
        ],
        "title": "ErrorCode",
        "type": "string"
      },
      "ExternalConversationData": {
        "additionalProperties": false,
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "external_conversation_id": {
            "title": "External Conversation Id",
            "type": "string"
          },
          "head_seq": {
            "pattern": "^(0|[1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
            "title": "Head Seq",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "conversation_id",
          "connection_id",
          "profile",
          "external_conversation_id",
          "title",
          "head_seq"
        ],
        "title": "ExternalConversationData",
        "type": "object"
      },
      "ExternalHistoryResponse": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ExternalMessageData"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/HistoryMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "ExternalHistoryResponse",
        "type": "object"
      },
      "ExternalMessageData": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Message Id"
          },
          "conversation_id": {
            "format": "uuid",
            "title": "Conversation Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "delivery_state": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DeliveryState"
              },
              {
                "type": "null"
              }
            ]
          },
          "effect_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effect Id"
          },
          "external_message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Message Id"
          },
          "external_sender_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Sender Id"
          },
          "message_id": {
            "format": "uuid",
            "title": "Message Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "operation_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operation Id"
          },
          "received_at": {
            "format": "date-time",
            "title": "Received At",
            "type": "string"
          },
          "sender_id": {
            "title": "Sender Id",
            "type": "string"
          },
          "sender_kind": {
            "$ref": "#/components/schemas/SenderKind"
          },
          "seq": {
            "pattern": "^([1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
            "title": "Seq",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "conversation_id",
          "sender_kind",
          "sender_id",
          "client_message_id",
          "external_message_id",
          "seq",
          "text",
          "occurred_at",
          "received_at",
          "created_at",
          "operation_id",
          "delivery_state",
          "external_sender_id",
          "effect_id"
        ],
        "title": "ExternalMessageData",
        "type": "object"
      },
      "FieldError": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/FieldErrorCode"
          },
          "location": {
            "items": {
              "type": "string"
            },
            "title": "Location",
            "type": "array"
          }
        },
        "required": [
          "location",
          "code"
        ],
        "title": "FieldError",
        "type": "object"
      },
      "FieldErrorCode": {
        "enum": [
          "REQUIRED",
          "TOO_SHORT",
          "TOO_LONG",
          "INVALID"
        ],
        "title": "FieldErrorCode",
        "type": "string"
      },
      "HistoryMeta": {
        "additionalProperties": false,
        "properties": {
          "has_more": {
            "title": "Has More",
            "type": "boolean"
          },
          "next_cursor": {
            "pattern": "^(0|[1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
            "title": "Next Cursor",
            "type": "string"
          },
          "snapshot_head_seq": {
            "pattern": "^(0|[1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
            "title": "Snapshot Head Seq",
            "type": "string"
          }
        },
        "required": [
          "next_cursor",
          "snapshot_head_seq",
          "has_more"
        ],
        "title": "HistoryMeta",
        "type": "object"
      },
...
      "InboundEvent": {
        "additionalProperties": false,
        "properties": {
          "connection_id": {
            "format": "uuid",
            "title": "Connection Id",
            "type": "string"
          },
          "external_conversation_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Conversation Id",
            "type": "string"
          },
          "external_event_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Event Id",
            "type": "string"
          },
          "external_message_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Message Id",
            "type": "string"
          },
          "external_sender_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "External Sender Id",
            "type": "string"
          },
          "occurred_at": {
            "format": "date-time",
            "title": "Occurred At",
            "type": "string"
          },
          "profile": {
            "$ref": "#/components/schemas/Profile"
          },
          "run_id": {
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9][A-Za-z0-9._:-]*$",
            "title": "Run Id",
            "type": "string"
          },
          "schema_version": {
            "const": 1,
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "run_id",
          "profile",
          "connection_id",
          "external_conversation_id",
          "external_event_id",
          "external_message_id",
          "external_sender_id",
          "occurred_at",
          "text"
        ],
        "title": "InboundEvent",
        "type": "object"
      },
      "Problem": {
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FieldError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "status": {
            "maximum": 599.0,
            "minimum": 400.0,
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "const": "about:blank",
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status",
          "code",
          "request_id"
        ],
        "title": "Problem",
        "type": "object"
      },
      "Profile": {
        "enum": [
          "telegram",
          "line",
          "instagram",
          "facebook",
          "whatsapp",
          "wechat",
          "kakao-bizgo"
        ],
        "title": "Profile",
        "type": "string"
      },
      "SendMessageRequest": {
        "additionalProperties": false,
        "properties": {
          "client_message_id": {
            "format": "uuid",
            "title": "Client Message Id",
            "type": "string"
          },
          "text": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "client_message_id",
          "text"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
      "SenderKind": {
        "enum": [
          "customer",
          "operator"
        ],
        "title": "SenderKind",
        "type": "string"
      },
...
      "Success_ExternalMessageData_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ExternalMessageData"
          }
        },
        "required": [
          "data"
        ],
        "title": "Success[ExternalMessageData]",
        "type": "object"
      },
...
      "Success_dict_str__UUID__": {
        "properties": {
          "data": {
            "additionalProperties": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Data",
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "title": "Success[dict[str, UUID]]",
        "type": "object"
      },
...
      "Success_list_ExternalConversationData__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/ExternalConversationData"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "Success[list[ExternalConversationData]]",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Laughtale Chat",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/dev/external-connections": {
      "post": {
        "operationId": "seed_connection_v1_dev_external_connections_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionSeed"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_dict_str__UUID__"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Seed Connection"
      }
    },
...
    "/v1/external-conversations": {
      "get": {
        "operationId": "external_rooms_v1_external_conversations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_list_ExternalConversationData__"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "External Rooms"
      }
    },
    "/v1/external-conversations/{conversation_id}/messages": {
      "get": {
        "operationId": "external_history_v1_external_conversations__conversation_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "after_seq",
            "required": false,
            "schema": {
              "default": "0",
              "pattern": "^(0|[1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
              "title": "After Seq",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "snapshot_head_seq",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "pattern": "^(0|[1-9][0-9]{0,17}|9223372036854775807|[1-8][0-9]{18}|9[0-1][0-9]{17}|92[0-1][0-9]{16}|922[0-2][0-9]{15}|9223[0-2][0-9]{14}|92233[0-6][0-9]{13}|922337[0-1][0-9]{12}|92233720[0-2][0-9]{10}|922337203[0-5][0-9]{9}|9223372036[0-7][0-9]{8}|92233720368[0-4][0-9]{7}|922337203685[0-3][0-9]{6}|9223372036854[0-6][0-9]{5}|92233720368547[0-6][0-9]{4}|922337203685477[0-4][0-9]{3}|9223372036854775[0-7][0-9]{2}|922337203685477580[0-6][0-9]{0})$",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Snapshot Head Seq"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalHistoryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "External History"
      },
      "post": {
        "operationId": "send_external_v1_external_conversations__conversation_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_ExternalMessageData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Send External"
      }
    },
    "/v1/external-conversations/{conversation_id}/messages/{message_id}": {
      "get": {
        "operationId": "external_message_v1_external_conversations__conversation_id__messages__message_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "conversation_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Conversation Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "message_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Message Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_ExternalMessageData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "External Message"
      }
    },
    "/v1/external-events": {
      "post": {
        "operationId": "receive_event_v1_external_events_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InboundEvent"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success_ExternalMessageData_"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not Found"
          },
          "405": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Method Not Allowed"
          },
          "422": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Unprocessable Content"
          },
          "500": {
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "summary": "Receive Event"
      }
    },
...
  }
}